Migrate back suggest to APIV1 (#2696)
