+++ js/components/organization/card-filter.vue
@@ -43,7 +43,7 @@
 </template>
 
 <script>
-import APIV2 from 'apiv2';
+import API from 'api';
 import log from 'logger';
 import Organization from 'models/organization';
 
@@ -84,7 +84,7 @@ export default {
     },
     watch: {
         search_query: function(query) {
-            APIV2.organizations.suggest_organizations({
+            API.organizations.suggest_organizations({
                 q: query,
                 size: 9
             }, (data) => {
