+++ udata/core/reuse/api_fields.py
@@ -77,11 +77,25 @@ reuse_ref_fields = api.inherit('ReuseReference', base_reference, {
         description='The reuse page URL', readonly=True, fallback_endpoint='api.reuse'),
 })
 
+
 reuse_type_fields = api.model('ReuseType', {
     'id': fields.String(description='The reuse type identifier'),
     'label': fields.String(description='The reuse type display name')
 })
 
+
+reuse_suggestion_fields = api.model('ReuseSuggestion', {
+    'id': fields.String(description='The reuse identifier', readonly=True),
+    'title': fields.String(description='The reuse title', readonly=True),
+    'slug': fields.String(
+        description='The reuse permalink string', readonly=True),
+    'image_url': fields.String(description='The reuse thumbnail URL'),
+    'page': fields.UrlFor(
+        'reuses.show_redirect', lambda o: {'reuse': o['slug']},
+        description='The reuse page URL', readonly=True, fallback_endpoint='api.reuse')
+})
+
+  
 reuse_topic_fields = api.model('ReuseTopic', {
     'id': fields.String(description='The reuse topic identifier'),
     'label': fields.String(description='The reuse topic display name')
