+++ udata/core/spatial/api_fields.py
@@ -66,3 +66,14 @@ spatial_coverage_fields = api.model('SpatialCoverage', {
     'granularity': fields.String(default='other',
         description='The spatial/territorial granularity'),
 })
+
+
+zone_suggestion_fields = api.model('TerritorySuggestion', {
+    'id': fields.String(description='The territory identifier', required=True),
+    'name': fields.String(description='The territory name', required=True),
+    'code': fields.String(
+        description='The territory main code', required=True),
+    'level': fields.String(
+        description='The territory administrative level', required=True),
+    'keys': fields.Raw(description='The territory known codes')
+})
