+++ udata/core/tags/api.py
@@ -1,13 +1,13 @@
-from udata.api import apiv2, API
+from udata.api import api, API
 
 from udata.tags import normalize  # TODO: merge this into this package
 from udata.models import Tag
 
 DEFAULT_SIZE = 8
 
-ns = apiv2.namespace('tags', 'Tags related operations')
+ns = api.namespace('tags', 'Tags related operations')
 
-parser = apiv2.parser()
+parser = api.parser()
 parser.add_argument(
     'q', type=str, help='The string to autocomplete/suggest',
     location='args', required=True)
@@ -17,9 +17,9 @@ parser.add_argument(
 
 
 @ns.route('/suggest/', endpoint='suggest_tags')
-class SuggestTagsAPIv2(API):
-    @apiv2.doc('suggest_tags')
-    @apiv2.expect(parser)
+class SuggestTagsAPI(API):
+    @api.doc('suggest_tags')
+    @api.expect(parser)
     def get(self):
         '''Suggest tags'''
         args = parser.parse_args()
