+++ udata/tests/api/test_datasets_api.py
@@ -995,7 +995,7 @@ class DatasetResourceAPITest(APITestCase):
             for f in (faker.word(), faker.word(), 'kml', 'kml-1')
         ])
 
-        response = self.get(url_for('apiv2.suggest_formats'),
+        response = self.get(url_for('api.suggest_formats'),
                             qs={'q': 'km', 'size': '5'})
         self.assert200(response)
 
@@ -1012,7 +1012,7 @@ class DatasetResourceAPITest(APITestCase):
             ResourceFactory(format=faker.word()) for _ in range(3)
         ])
 
-        response = self.get(url_for('apiv2.suggest_formats'),
+        response = self.get(url_for('api.suggest_formats'),
                             qs={'q': 'test', 'size': '5'})
         self.assert200(response)
         self.assertEqual(len(response.json), 0)
@@ -1020,7 +1020,7 @@ class DatasetResourceAPITest(APITestCase):
     def test_suggest_format_api_empty(self):
         '''It should not provide format suggestion if no data'''
         self.init_search()
-        response = self.get(url_for('apiv2.suggest_formats'),
+        response = self.get(url_for('api.suggest_formats'),
                             qs={'q': 'test', 'size': '5'})
         self.assert200(response)
         self.assertEqual(len(response.json), 0)
@@ -1036,7 +1036,7 @@ class DatasetResourceAPITest(APITestCase):
             )
         ])
 
-        response = self.get(url_for('apiv2.suggest_mime'),
+        response = self.get(url_for('api.suggest_mime'),
                             qs={'q': 'js', 'size': '5'})
         self.assert200(response)
         self.assertLessEqual(len(response.json), 5)
@@ -1048,7 +1048,7 @@ class DatasetResourceAPITest(APITestCase):
         '''It should suggest mime types'''
         DatasetFactory(resources=[ResourceFactory(mime='application/xhtml+xml')])
 
-        response = self.get(url_for('apiv2.suggest_mime'),
+        response = self.get(url_for('api.suggest_mime'),
                             qs={'q': 'xml', 'size': '5'})
         self.assert200(response)
 
@@ -1060,7 +1060,7 @@ class DatasetResourceAPITest(APITestCase):
             ResourceFactory(mime=faker.word()) for _ in range(3)
         ])
 
-        response = self.get(url_for('apiv2.suggest_mime'),
+        response = self.get(url_for('api.suggest_mime'),
                             qs={'q': 'test', 'size': '5'})
         self.assert200(response)
         self.assertEqual(len(response.json), 0)
@@ -1068,7 +1068,7 @@ class DatasetResourceAPITest(APITestCase):
     def test_suggest_mime_api_empty(self):
         '''It should not provide mime suggestion if no data'''
         self.init_search()
-        response = self.get(url_for('apiv2.suggest_mime'),
+        response = self.get(url_for('api.suggest_mime'),
                             qs={'q': 'test', 'size': '5'})
         self.assert200(response)
         self.assertEqual(len(response.json), 0)
@@ -1080,7 +1080,7 @@ class DatasetResourceAPITest(APITestCase):
                 title='test-{0}'.format(i) if i % 2 else faker.word(),
                 visible=True)
 
-        response = self.get(url_for('apiv2.suggest_datasets'),
+        response = self.get(url_for('api.suggest_datasets'),
                             qs={'q': 'tes', 'size': '5'})
         self.assert200(response)
 
@@ -1103,7 +1103,7 @@ class DatasetResourceAPITest(APITestCase):
                 acronym='test-{0}'.format(i) if i % 2 else None,
                 visible=True)
 
-        response = self.get(url_for('apiv2.suggest_datasets'),
+        response = self.get(url_for('api.suggest_datasets'),
                             qs={'q': 'tes', 'size': '5'})
         self.assert200(response)
 
@@ -1125,7 +1125,7 @@ class DatasetResourceAPITest(APITestCase):
                 title='testé-{0}'.format(i) if i % 2 else faker.word(),
                 resources=[ResourceFactory()])
 
-        response = self.get(url_for('apiv2.suggest_datasets'),
+        response = self.get(url_for('api.suggest_datasets'),
                             qs={'q': 'testé', 'size': '5'})
         self.assert200(response)
 
@@ -1144,7 +1144,7 @@ class DatasetResourceAPITest(APITestCase):
         for i in range(3):
             DatasetFactory(resources=[ResourceFactory()])
 
-        response = self.get(url_for('apiv2.suggest_datasets'),
+        response = self.get(url_for('api.suggest_datasets'),
                             qs={'q': 'xxxxxx', 'size': '5'})
         self.assert200(response)
         self.assertEqual(len(response.json), 0)
@@ -1152,7 +1152,7 @@ class DatasetResourceAPITest(APITestCase):
     def test_suggest_datasets_api_empty(self):
         '''It should not provide dataset suggestion if no data'''
         self.init_search()
-        response = self.get(url_for('apiv2.suggest_datasets'),
+        response = self.get(url_for('api.suggest_datasets'),
                             qs={'q': 'xxxxxx', 'size': '5'})
         self.assert200(response)
         self.assertEqual(len(response.json), 0)
