+++ udata/tests/api/test_reuses_api.py
@@ -263,7 +263,7 @@ class ReuseAPITest:
                 title='test-{0}'.format(i) if i % 2 else faker.word(),
                 visible=True)
 
-        response = api.get(url_for('apiv2.suggest_reuses'),
+        response = api.get(url_for('api.suggest_reuses'),
                            qs={'q': 'tes', 'size': '5'})
         assert200(response)
 
@@ -284,7 +284,7 @@ class ReuseAPITest:
                 title='testé-{0}'.format(i) if i % 2 else faker.word(),
                 visible=True)
 
-        response = api.get(url_for('apiv2.suggest_reuses'),
+        response = api.get(url_for('api.suggest_reuses'),
                            qs={'q': 'testé', 'size': '5'})
         assert200(response)
 
@@ -302,7 +302,7 @@ class ReuseAPITest:
         '''It should not provide reuse suggestion if no match'''
         ReuseFactory.create_batch(3, visible=True)
 
-        response = api.get(url_for('apiv2.suggest_reuses'),
+        response = api.get(url_for('api.suggest_reuses'),
                            qs={'q': 'xxxxxx', 'size': '5'})
         assert200(response)
         assert len(response.json) == 0
@@ -310,7 +310,7 @@ class ReuseAPITest:
     def test_suggest_reuses_api_empty(self, api):
         '''It should not provide reuse suggestion if no data'''
         # self.init_search()
-        response = api.get(url_for('apiv2.suggest_reuses'),
+        response = api.get(url_for('api.suggest_reuses'),
                            qs={'q': 'xxxxxx', 'size': '5'})
         assert200(response)
         assert len(response.json) == 0
