+++ udata/tests/api/test_tags_api.py
@@ -21,7 +21,7 @@ class TagsAPITest:
 
         count_tags()
 
-        response = api.get(url_for('apiv2.suggest_tags'),
+        response = api.get(url_for('api.suggest_tags'),
                             qs={'q': 'tes', 'size': '5'})
         assert200(response)
 
@@ -43,7 +43,7 @@ class TagsAPITest:
 
         count_tags()
 
-        response = api.get(url_for('apiv2.suggest_tags'),
+        response = api.get(url_for('api.suggest_tags'),
                            qs={'q': 'testé', 'size': '5'})
         assert200(response)
 
@@ -64,14 +64,14 @@ class TagsAPITest:
 
         count_tags()
 
-        response = api.get(url_for('apiv2.suggest_tags'),
+        response = api.get(url_for('api.suggest_tags'),
                            qs={'q': 'bbbb', 'size': '5'})
         assert200(response)
         assert len(response.json) is 0
 
     def test_suggest_tags_api_empty(self, api, autoindex):
         '''It should not provide tag suggestion if no data'''
-        response = api.get(url_for('apiv2.suggest_tags'),
+        response = api.get(url_for('api.suggest_tags'),
                            qs={'q': 'bbbb', 'size': '5'})
         assert200(response)
         assert len(response.json) is 0
